pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-bot"
version = "0.1.0"
description = "Alpaca-based algorithmic trading scaffold with backtesting, paper trading, and monitoring."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alpaca-py>=0.42.0",
  "numpy>=1.26",
  "pandas>=2.2",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "pytz>=2024.1",
  "python-dotenv>=1.0",
  "SQLAlchemy>=2.0",
  "streamlit>=1.44",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10",
  "pytest>=8.2",
  "ruff>=0.5",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
  "ignore:websockets\\.legacy is deprecated.*:DeprecationWarning",
]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
